refactor(utils): ♻️ merge schema helpers
