+++ src/utils/extract-schemas.ts
@@ -1,5 +1,6 @@
 import type { OpenAPIV3_1 as OpenAPI } from 'openapi-types';
-import { collectSchemaRefs } from './filter-paths.js';
+import { collectSchemaRefs } from './collect-schema-refs.js';
+import { collectSchemas } from './collect-schemas.js';
 
 export function extractSchemas(
   doc: OpenAPI.Document,
@@ -25,25 +26,5 @@ export function extractSchemas(
     return {};
   }
 
-  const collected: Record<string, OpenAPI.SchemaObject | OpenAPI.ReferenceObject> = {};
-  const queue = Array.from(refs);
-  const processed = new Set<string>();
-
-  while (queue.length) {
-    const name = queue.pop()!;
-    if (processed.has(name)) continue;
-    const schema = allSchemas[name];
-    if (schema) {
-      collected[name] = schema;
-      processed.add(name);
-      collectSchemaRefs(schema, refs);
-      for (const ref of refs) {
-        if (!processed.has(ref) && !queue.includes(ref)) {
-          queue.push(ref);
-        }
-      }
-    }
-  }
-
-  return collected;
+  return collectSchemas(doc, refs);
 }
